LARA answer conversion script: move source dir not to inferfere with the main repo build process
[#182525578]
