+++ convert-lara-answers/config-sample.json
@@ -4,5 +4,5 @@
   "newSourceKey": "activity-player.concord.org",
   "startDate": "2020-01-01 12:00:00 UTC",
   "endDate": "2020-01-10 12:00:00 UTC",
-  "maxDocCount": "[temporary - TODO: don't use during the final conversion]"
+  "maxDocCount": 10e20
 }
