+++ convert-lara-answers/src/conversion-helpers/convert-image-question.test.ts
