+++ convert-lara-answers/src/conversion-helpers/convert-image-question.ts
@@ -1,4 +1,4 @@
-import { IInteractiveState as IImageQuestionInteractiveState } from "../../../image-question/components/types";
+import { IInteractiveState as IImageQuestionInteractiveState } from "../../../src/image-question/components/types";
 import { convertAnswer } from "./convert-answer";
 import { ILARAAnonymousAnswerReportHash, ILARAAnswerReportHash, ILARAImageQuestionAnswerReportHash, IManagedInteractiveQuestion } from "../types";
 import { getReportState } from "../utils";
