+++ convert-lara-answers/src/conversion-helpers/convert-interactive-state.ts
