+++ convert-lara-answers/src/log-utils.ts
