+++ convert-lara-answers/tsconfig.json
@@ -2,8 +2,6 @@
   "$schema": "https://json.schemastore.org/tsconfig",
   "display": "Node 16",
 
-
-
   "compilerOptions": {
     "outDir": "./dist",
     "sourceMap": true,
